+++ src/jarvis/jarvis/sensor/image_data.h
@@ -14,7 +14,7 @@ struct ImageData {
     std::vector<std::vector<cv::Mat>> result;
     for (size_t i = 0; i < pyramid_derive.size(); i++) {
       result.emplace_back();
-      for (int j = 0; j < pyramid_derive[i].size(); j += 2)
+      for (size_t j = 0; j < pyramid_derive[i].size(); j += 2)
         result.back().push_back(pyramid_derive[i][j]);
     }
     return result;
